Moved bindings

+++ runtime/bindings/c/samples/hello_nv12_input_classification/main.c
@@ -103,9 +103,9 @@ void print_classify_res(struct classify_res* cls, size_t n, const char* img_path
  * @return total number of elements successfully read, in case of error it
  * doesn't equal to size param
  */
-size_t read_image_from_file(const char* img_path, unsigned char* img_data, size_t size) {
+long int read_image_from_file(const char* img_path, unsigned char* img_data, long int size) {
     FILE* fp = fopen(img_path, "rb+");
-    size_t read_size = 0;
+    long int read_size = 0;
 
     if (fp) {
         fseek(fp, 0, SEEK_END);
